Backend/requirements.txt
ipython
google-generativeai
qdrant-client
fastapi
uvicorn[standard]
beautifulsoup4
numpy
jinja2
google-auth-oauthlib